@@ -7,29 +7,32 @@
         "psr-4": { "": "src/" },
         "classmap": [ "app/AppKernel.php", "app/AppCache.php" ]
     },
+    "autoload-dev": {
+        "psr-4": { "Tests\\": "tests/" }
+    },
     "require": {
-        "php"                                  : ">=5.3.9",
+        "php"                                  : ">=5.5.9",
         "ext-pdo_sqlite"                       : "*",
-        "doctrine/doctrine-bundle"             : "~1.5",
-        "doctrine/doctrine-fixtures-bundle"    : "~2.2",
-        "doctrine/orm"                         : "~2.4",
-        "erusev/parsedown"                     : "~1.5",
-        "ezyang/htmlpurifier"                  : "~4.7",
-        "incenteev/composer-parameter-handler" : "~2.1",
-        "ircmaxell/password-compat"            : "~1.0",
-        "knplabs/knp-paginator-bundle"         : "~2.4",
-        "leafo/scssphp"                        : "~0.1.5",
-        "patchwork/jsqueeze"                   : "~1.0",
-        "sensio/distribution-bundle"           : "~5.0",
-        "sensio/framework-extra-bundle"        : "~3.0",
-        "symfony/assetic-bundle"               : "~2.6",
-        "symfony/monolog-bundle"               : "~2.7",
-        "symfony/swiftmailer-bundle"           : "~2.3",
-        "symfony/symfony"                      : "~2.8",
-        "twig/extensions"                      : "~1.2"
+        "doctrine/doctrine-bundle"             : "^1.6",
+        "doctrine/doctrine-fixtures-bundle"    : "^2.2",
+        "doctrine/orm"                         : "^2.5",
+        "erusev/parsedown"                     : "^1.5",
+        "ezyang/htmlpurifier"                  : "^4.7",
+        "incenteev/composer-parameter-handler" : "^2.0",
+        "leafo/scssphp"                        : "^0.5",
+        "patchwork/jsqueeze"                   : "^2.0",
+        "sensio/distribution-bundle"           : "^5.0",
+        "sensio/framework-extra-bundle"        : "^3.0",
+        "symfony/assetic-bundle"               : "^2.8",
+        "symfony/monolog-bundle"               : "^2.8",
+        "symfony/swiftmailer-bundle"           : "^2.3",
+        "symfony/symfony"                      : "^3.1",
+        "twig/extensions"                      : "^1.3",
+        "white-october/pagerfanta-bundle"      : "^1.0"
     },
     "require-dev": {
-        "sensio/generator-bundle": "~3.0"
+        "sensio/generator-bundle"              : "~3.0",
+        "symfony/phpunit-bridge"               : "^3.0"
     },
     "scripts": {
         "post-install-cmd": [
@@ -52,15 +55,22 @@
     "config": {
         "bin-dir": "bin",
         "platform": {
-            "php": "5.3.9"
+            "php": "5.5.9"
         }
     },
     "extra": {
         "symfony-app-dir": "app",
+        "symfony-bin-dir": "bin",
+        "symfony-var-dir": "var",
         "symfony-web-dir": "web",
+        "symfony-tests-dir": "tests",
         "symfony-assets-install": "relative",
         "incenteev-parameters": {
-            "file": "app/config/parameters.yml"
+            "file": "app/config/parameters.yml",
+            "env-map": {
+                "database_url": "DATABASE_URL",
+                "secret": "SYMFONY_SECRET"
+            }
         }
     }
 }
