@@ -11,12 +11,14 @@
 
 namespace AppBundle\Controller\Admin;
 
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use AppBundle\Entity\Post;
+use AppBundle\Form\PostType;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
-use AppBundle\Entity\Post;
+use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Symfony\Component\Form\Extension\Core\Type\SubmitType;
+use Symfony\Component\HttpFoundation\Request;
 
 /**
  * Controller used to manage blog contents in the backend.
@@ -52,9 +54,9 @@ class BlogController extends Controller
     public function indexAction()
     {
         $entityManager = $this->getDoctrine()->getManager();
-        $posts = $entityManager->getRepository('AppBundle:Post')->findAll();
+        $posts = $entityManager->getRepository(Post::class)->findAll();
 
-        return $this->render('admin/blog/index.html.twig', array('posts' => $posts));
+        return $this->render('admin/blog/index.html.twig', ['posts' => $posts]);
     }
 
     /**
@@ -73,8 +75,8 @@ class BlogController extends Controller
         $post->setAuthorEmail($this->getUser()->getEmail());
 
         // See http://symfony.com/doc/current/book/forms.html#submitting-forms-with-multiple-buttons
-        $form = $this->createForm('AppBundle\Form\PostType', $post)
-            ->add('saveAndCreateNew', 'Symfony\Component\Form\Extension\Core\Type\SubmitType');
+        $form = $this->createForm(PostType::class, $post)
+            ->add('saveAndCreateNew', SubmitType::class);
 
         $form->handleRequest($request);
 
@@ -102,16 +104,16 @@ class BlogController extends Controller
             return $this->redirectToRoute('admin_post_index');
         }
 
-        return $this->render('admin/blog/new.html.twig', array(
+        return $this->render('admin/blog/new.html.twig', [
             'post' => $post,
             'form' => $form->createView(),
-        ));
+        ]);
     }
 
     /**
      * Finds and displays a Post entity.
      *
-     * @Route("/{id}", requirements={"id" = "\d+"}, name="admin_post_show")
+     * @Route("/{id}", requirements={"id": "\d+"}, name="admin_post_show")
      * @Method("GET")
      */
     public function showAction(Post $post)
@@ -125,16 +127,16 @@ class BlogController extends Controller
 
         $deleteForm = $this->createDeleteForm($post);
 
-        return $this->render('admin/blog/show.html.twig', array(
+        return $this->render('admin/blog/show.html.twig', [
             'post'        => $post,
             'delete_form' => $deleteForm->createView(),
-        ));
+        ]);
     }
 
     /**
      * Displays a form to edit an existing Post entity.
      *
-     * @Route("/{id}/edit", requirements={"id" = "\d+"}, name="admin_post_edit")
+     * @Route("/{id}/edit", requirements={"id": "\d+"}, name="admin_post_edit")
      * @Method({"GET", "POST"})
      */
     public function editAction(Post $post, Request $request)
@@ -145,7 +147,7 @@ class BlogController extends Controller
 
         $entityManager = $this->getDoctrine()->getManager();
 
-        $editForm = $this->createForm('AppBundle\Form\PostType', $post);
+        $editForm = $this->createForm(PostType::class, $post);
         $deleteForm = $this->createDeleteForm($post);
 
         $editForm->handleRequest($request);
@@ -156,14 +158,14 @@ class BlogController extends Controller
 
             $this->addFlash('success', 'post.updated_successfully');
 
-            return $this->redirectToRoute('admin_post_edit', array('id' => $post->getId()));
+            return $this->redirectToRoute('admin_post_edit', ['id' => $post->getId()]);
         }
 
-        return $this->render('admin/blog/edit.html.twig', array(
+        return $this->render('admin/blog/edit.html.twig', [
             'post'        => $post,
             'edit_form'   => $editForm->createView(),
             'delete_form' => $deleteForm->createView(),
-        ));
+        ]);
     }
 
     /**
@@ -210,7 +212,7 @@ class BlogController extends Controller
     private function createDeleteForm(Post $post)
     {
         return $this->createFormBuilder()
-            ->setAction($this->generateUrl('admin_post_delete', array('id' => $post->getId())))
+            ->setAction($this->generateUrl('admin_post_delete', ['id' => $post->getId()]))
             ->setMethod('DELETE')
             ->getForm()
         ;
