@@ -40,17 +40,17 @@ class SourceCodeExtension extends \Twig_Extension
      */
     public function getFunctions()
     {
-        return array(
-            new \Twig_SimpleFunction('show_source_code', array($this, 'showSourceCode'), array('is_safe' => array('html'), 'needs_environment' => true)),
-        );
+        return [
+            new \Twig_SimpleFunction('show_source_code', [$this, 'showSourceCode'], ['is_safe' => ['html'], 'needs_environment' => true]),
+        ];
     }
 
-    public function showSourceCode(\Twig_Environment $twig, \Twig_Template $template)
+    public function showSourceCode(\Twig_Environment $twig, $template)
     {
-        return $twig->render('@CodeExplorer/source_code.html.twig', array(
+        return $twig->render('@CodeExplorer/source_code.html.twig', [
             'controller' => $this->getController(),
-            'template'   => $this->getTemplateSource($template),
-        ));
+            'template'   => $this->getTemplateSource($twig->resolveTemplate($template)),
+        ]);
     }
 
     private function getController()
@@ -62,15 +62,15 @@ class SourceCodeExtension extends \Twig_Extension
 
         $method = $this->getCallableReflector($this->controller);
 
-        $classCode = file($method->getFilename());
-        $methodCode = array_slice($classCode, $method->getStartline() - 1, $method->getEndLine() - $method->getStartline() + 1);
+        $classCode = file($method->getFileName());
+        $methodCode = array_slice($classCode, $method->getStartLine() - 1, $method->getEndLine() - $method->getStartLine() + 1);
         $controllerCode = '    '.$method->getDocComment()."\n".implode('', $methodCode);
 
-        return array(
-            'file_path' => $method->getFilename(),
-            'starting_line' => $method->getStartline(),
+        return [
+            'file_path' => $method->getFileName(),
+            'starting_line' => $method->getStartLine(),
             'source_code' => $this->unindentCode($controllerCode)
-        );
+        ];
     }
 
     /**
@@ -99,16 +99,24 @@ class SourceCodeExtension extends \Twig_Extension
 
     private function getTemplateSource(\Twig_Template $template)
     {
-        return array(
-            // Twig templates are not always stored in files, and so there is no
-            // API to get the filename from a template name in a generic way.
-            // The logic used here works only for templates stored in app/Resources/views
-            // and referenced via the "filename.html.twig" notation, not via the "::filename.html.twig"
-            // one or stored in bundles. This is enough for the needs of the demo app.
-            'file_path' => $this->kernelRootDir.'/Resources/views/'.$template->getTemplateName(),
+        // Twig templates are not always stored in files, and so there is no
+        // API to get the filename from a template name in a generic way.
+        // The logic used here works only for templates stored in app/Resources/views
+        // and referenced via the "filename.html.twig" notation, not via the "::filename.html.twig"
+        // one or stored in bundles. This is enough for the needs of the demo app.
+        $filePath = $this->kernelRootDir.'/Resources/views/'.$template->getTemplateName();
+        $sourceCode = $template->getSource();
+
+        // Temporary workaround for https://github.com/twigphp/Twig/issues/2011
+        if (null === $sourceCode) {
+            $sourceCode = @file_get_contents($filePath);
+        }
+
+        return [
+            'file_path' => $filePath,
             'starting_line' => 1,
-            'source_code' => $template->getSource(),
-        );
+            'source_code' => $sourceCode,
+        ];
     }
 
     /**
