@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Tests\Utils;
+namespace Tests\AppBundle\Utils;
 
 use AppBundle\Utils\Slugger;
 
@@ -38,12 +38,10 @@ class SluggerTest extends \PHPUnit_Framework_TestCase
 
     public function getSlugs()
     {
-        return array(
-            array('Lorem Ipsum'     , 'lorem-ipsum'),
-            array('  Lorem Ipsum  ' , 'lorem-ipsum'),
-            array(' lOrEm  iPsUm  ' , 'lorem-ipsum'),
-            array('!Lorem Ipsum!'   , 'lorem-ipsum'),
-            array('lorem-ipsum'     , 'lorem-ipsum'),
-        );
+        yield ['Lorem Ipsum'     , 'lorem-ipsum'];
+        yield ['  Lorem Ipsum  ' , 'lorem-ipsum'];
+        yield [' lOrEm  iPsUm  ' , 'lorem-ipsum'];
+        yield ['!Lorem Ipsum!'   , 'lorem-ipsum'];
+        yield ['lorem-ipsum'     , 'lorem-ipsum'];
     }
 }
