@@ -12,7 +12,7 @@ use Symfony\Component\HttpFoundation\Request;
  * @var Composer\Autoload\ClassLoader
  */
 $loader = require __DIR__.'/../app/autoload.php';
-include_once __DIR__.'/../app/bootstrap.php.cache';
+include_once __DIR__.'/../var/bootstrap.php.cache';
 
 // If your web server provides APC support for PHP applications, uncomment these
 // lines to use APC for class autoloading. This can improve application performance
@@ -38,7 +38,7 @@ $kernel = new AppCache($kernel);
 // If you use HTTP Cache and your application relies on the _method request parameter
 // to get the intended HTTP method, uncomment this line.
 // See http://symfony.com/doc/current/reference/configuration/framework.html#http-method-override
-// Request::enableHttpMethodParameterOverride();
+Request::enableHttpMethodParameterOverride();
 
 $request = Request::createFromGlobals();
 $response = $kernel->handle($request);
