@@ -1,15 +1,28 @@
 <?php
 
+/*
+ * ************** CAUTION **************
+ *
+ * DO NOT EDIT THIS FILE as it will be overridden by Composer as part of
+ * the installation/update process. The original file resides in the
+ * SensioDistributionBundle.
+ *
+ * ************** CAUTION **************
+ */
+
 if (!isset($_SERVER['HTTP_HOST'])) {
-    exit('This script cannot be run from the console. Run it from a browser.');
+    exit('This script cannot be run from the CLI. Run it from a browser.');
 }
 
-if (!in_array(@$_SERVER['REMOTE_ADDR'], array('127.0.0.1', '::1'))) {
+if (!in_array(@$_SERVER['REMOTE_ADDR'], array(
+    '127.0.0.1',
+    '::1',
+))) {
     header('HTTP/1.0 403 Forbidden');
     exit('This script is only accessible from localhost.');
 }
 
-require_once dirname(__FILE__).'/../app/SymfonyRequirements.php';
+require_once dirname(__FILE__).'/../var/SymfonyRequirements.php';
 
 $symfonyRequirements = new SymfonyRequirements();
 
@@ -17,77 +30,157 @@ $majorProblems = $symfonyRequirements->getFailedRequirements();
 $minorProblems = $symfonyRequirements->getFailedRecommendations();
 
 ?>
-
 <!DOCTYPE html>
 <html>
     <head>
-        <meta charset="UTF-8" />
-        <title>Symfony Demo Application</title>
-        <link rel="stylesheet" href="css/app.css">
-        <link rel="icon" type="image/x-icon" href="favicon.ico" />
+        <meta http-equiv="Content-Type" content="text/html; charset=UTF-8"/>
+        <meta name="robots" content="noindex,nofollow" />
+        <title>Symfony Configuration Checker</title>
+        <link rel="stylesheet" href="bundles/framework/css/structure.css" media="all" />
+        <link rel="stylesheet" href="bundles/framework/css/body.css" media="all" />
+        <style type="text/css">
+            /* styles copied from bundles/sensiodistribution/webconfigurator/css/install.css */
+            body {
+                font-size: 14px;
+                font-family: "Lucida Sans Unicode", "Lucida Grande", Verdana, Arial, Helvetica, sans-serif;
+            }
+            .sf-reset h1.title {
+                font-size: 45px;
+                padding-bottom: 30px;
+            }
+            .sf-reset h2 {
+                font-weight: bold;
+                color: #FFFFFF;
+                /* Font is reset to sans-serif (like body) */
+                font-family: "Lucida Sans Unicode", "Lucida Grande", Verdana, Arial, Helvetica, sans-serif;
+                margin-bottom: 10px;
+                background-color: #aacd4e;
+                padding: 2px 4px;
+                display: inline-block;
+                text-transform: uppercase;
+            }
+            .sf-reset ul a,
+            .sf-reset ul a:hover {
+                background: url(../images/blue-arrow.png) no-repeat right 6px;
+                padding-right: 10px;
+            }
+            .sf-reset ul, ol {
+                padding-left: 20px;
+            }
+            .sf-reset li {
+                padding-bottom: 18px;
+            }
+            .sf-reset ol li {
+                list-style-type: decimal;
+            }
+            .sf-reset ul li {
+                list-style-type: none;
+            }
+            .sf-reset .symfony-blocks-install {
+                overflow: hidden;
+            }
+            .sf-reset .symfony-install-continue {
+                font-size: 0.95em;
+                padding-left: 0;
+            }
+            .sf-reset .symfony-install-continue li {
+                padding-bottom: 10px;
+            }
+            .sf-reset .ok {
+                color: #fff;
+                font-family: "Lucida Sans Unicode", "Lucida Grande", Verdana, Arial, Helvetica, sans-serif;
+                background-color: #6d6;
+                padding: 10px;
+                margin-bottom: 20px;
+            }
+            .sf-reset .ko {
+                background-color: #d66;
+            }
+            .version {
+                text-align: right;
+                font-size: 10px;
+                margin-right: 20px;
+            }
+            .sf-reset a,
+            .sf-reset li a {
+                color: #08C;
+                text-decoration: none;
+            }
+            .sf-reset a:hover,
+            .sf-reset li a:hover {
+                color: #08C;
+                text-decoration: underline;
+            }
+            .sf-reset textarea {
+                padding: 7px;
+            }
+        </style>
     </head>
+    <body>
+        <div id="content">
+            <div class="header clear-fix">
+                <div class="header-logo">
+                    <img src="bundles/framework/images/logo_symfony.png" alt="Symfony" />
+                </div>
 
-    <body id="requirements_checker">
-        <header>
-            <div class="navbar navbar-default navbar-static-top" role="navigation">
-                <div class="container">
-                    <div class="navbar-header">
-                        <h1 class="navbar-brand">
-                            Symfony Demo Application <span>/</span> Technical Requirements Checker
-                        </h1>
+                <div class="search">
+                  <form method="get" action="http://symfony.com/search">
+                    <div class="form-row">
+
+                      <label for="search-id">
+                          <img src="bundles/framework/images/grey_magnifier.png" alt="Search on Symfony website" />
+                      </label>
+
+                      <input name="q" id="search-id" type="search" placeholder="Search on Symfony website" />
+
+                      <button type="submit" class="sf-button">
+                          <span class="border-l">
+                            <span class="border-r">
+                                <span class="btn-bg">OK</span>
+                            </span>
+                        </span>
+                      </button>
                     </div>
+                   </form>
                 </div>
             </div>
-        </header>
 
-        <div class="container body-container">
-            <div class="row">
-                <div id="main" class="col-sm-9">
-                    <div class="row">
+            <div class="sf-reset">
+                <div class="block">
+                    <div class="symfony-block-content">
+                        <h1 class="title">Configuration Checker</h1>
+                        <p>
+                            This script analyzes your system to check whether is
+                            ready to run Symfony applications.
+                        </p>
+
                         <?php if (count($majorProblems)): ?>
-                            <div class="panel panel-danger">
-                                <div class="panel-heading">
-                                    <h3 class="panel-title">
-                                        <i class="fa fa-exclamation-circle"></i>
-                                        Fix these problems before continuing
-                                    </h3>
-                                </div>
-                                <div class="panel-body">
-                                    <p>These <strong>mandatory requirements</strong> must be fixed before running Symfony applications:</p>
-
-                                    <ol>
-                                        <?php foreach ($majorProblems as $problem): ?>
-                                            <li><?php echo $problem->getHelpHtml() ?></li>
-                                        <?php endforeach; ?>
-                                    </ol>
-                                </div>
-                            </div>
+                            <h2 class="ko">Major problems</h2>
+                            <p>Major problems have been detected and <strong>must</strong> be fixed before continuing:</p>
+                            <ol>
+                                <?php foreach ($majorProblems as $problem): ?>
+                                    <li><?php echo $problem->getHelpHtml() ?></li>
+                                <?php endforeach; ?>
+                            </ol>
                         <?php endif; ?>
 
                         <?php if (count($minorProblems)): ?>
-                            <div class="panel panel-warning">
-                                <div class="panel-heading">
-                                    <h3 class="panel-title">
-                                        <i class="fa fa-exclamation-triangle"></i>
-                                        Fix these problems to improve your experience
-                                    </h3>
-                                </div>
-                                <div class="panel-body">
-                                    <p>These <strong>optional requirements</strong> should be fixed to improve your experience running Symfony applications:</p>
-
-                                    <ol>
-                                        <?php foreach ($minorProblems as $problem): ?>
-                                            <li><?php echo $problem->getHelpHtml() ?></li>
-                                        <?php endforeach; ?>
-                                    </ol>
-                                </div>
-                            </div>
+                            <h2>Recommendations</h2>
+                            <p>
+                                <?php if (count($majorProblems)): ?>Additionally, to<?php else: ?>To<?php endif; ?> enhance your Symfony experience,
+                                it’s recommended that you fix the following:
+                            </p>
+                            <ol>
+                                <?php foreach ($minorProblems as $problem): ?>
+                                    <li><?php echo $problem->getHelpHtml() ?></li>
+                                <?php endforeach; ?>
+                            </ol>
                         <?php endif; ?>
 
                         <?php if ($symfonyRequirements->hasPhpIniConfigIssue()): ?>
                             <p id="phpini">*
                                 <?php if ($symfonyRequirements->getPhpIniConfigPath()): ?>
-                                    Changes to the <strong>php.ini</strong> file must be done in the "<strong><?php echo $symfonyRequirements->getPhpIniConfigPath() ?></strong>" file.
+                                    Changes to the <strong>php.ini</strong> file must be done in "<strong><?php echo $symfonyRequirements->getPhpIniConfigPath() ?></strong>".
                                 <?php else: ?>
                                     To change settings, create a "<strong>php.ini</strong>".
                                 <?php endif; ?>
@@ -95,57 +188,18 @@ $minorProblems = $symfonyRequirements->getFailedRecommendations();
                         <?php endif; ?>
 
                         <?php if (!count($majorProblems) && !count($minorProblems)): ?>
-                            <div class="panel panel-success">
-                                <div class="panel-heading">
-                                    <h3 class="panel-title">
-                                        <i class="fa fa-check"></i>
-                                        All checks passed successfully
-                                    </h3>
-                                </div>
-                                <div class="panel-body">
-                                    <p>Your configuration looks good to run the Symfony Demo application.</p>
-                                </div>
-                            </div>
+                            <p class="ok">All checks passed successfully. Your system is ready to run Symfony applications.</p>
                         <?php endif; ?>
-                    </div>
-                </div>
 
-                <div id="sidebar" class="col-sm-3">
-                    <div class="section about">
-                        <div class="well well-lg">
-                            <p>
-                                This script checks whether your system meets the
-                                technical requirements for running Symfony
-                                applications.
-                            </p>
-                            <p>
-                                For more information, check out the
-                                <a href="http://symfony.com/doc/current/reference/requirements.html">requirements list</a>
-                                at the Symfony documentation.
-                            </p>
-                        </div>
+                        <ul class="symfony-install-continue">
+                            <?php if (count($majorProblems) || count($minorProblems)): ?>
+                                <li><a href="config.php">Re-check configuration</a></li>
+                            <?php endif; ?>
+                        </ul>
                     </div>
-
                 </div>
             </div>
+            <div class="version">Symfony Standard Edition</div>
         </div>
-
-        <footer>
-            <div class="container">
-                <div class="row">
-                    <div id="footer-copyright" class="col-md-6">
-                        <p>&copy; 2015 - The Symfony Project</p>
-                        <p>MIT License</p>
-                    </div>
-                    <div id="footer-resources" class="col-md-6">
-                        <p>
-                            <a href="https://twitter.com/symfony"><i class="fa fa-twitter"></i></a>
-                            <a href="https://www.facebook.com/SensioLabs"><i class="fa fa-facebook"></i></a>
-                            <a href="http://symfony.com/blog"><i class="fa fa-rss"></i></a>
-                        </p>
-                    </div>
-                </div>
-            </div>
-        </footer>
     </body>
 </html>
